# dispatch-log: use unbuffered channels, detect rotation via fs events