@samth 2022-06-13
The improvement to the distribution of response times seems quite significant. 6x better at p90!